Conditions

These functions provide a way for signalling conditions. rui::warn() and rui::error() are drop-in replacements for base::warning() and base::stop(), which support glue strings and cli inline styles. ANSI colours are lost however. For retaining colours, rui::alert() can be used. We recommend doing so before issuing a warning or error.

alert(..., warn = FALSE, error = FALSE)

warn(..., .demo = FALSE)

error(..., .demo = FALSE)

stop(...)

Arguments

...

Character vectors supporting glue strings and cli inline styles.

warn

Logical. Should a warning be issued with the same message.

error

Logical. Should an error be issued with the same message.

See also